feat(manager-components): rename to sshKeySelector component
ref: MANAGER-14906
