+++ packages/manager/modules/bm-server-components/src/ipmi/sol/add/sol-add-ssh.html
@@ -13,10 +13,10 @@
             ></h2>
 
             <form name="solSshSession">
-                <ssh-key
+                <ssh-key-selector
                     data-public-key="$ctrl.ssh.publicKey"
                     data-required="true"
-                ></ssh-key>
+                ></ssh-key-selector>
             </form>
         </div>
     </div>
