+++ packages/manager/modules/manager-components/src/index.js
@@ -5,7 +5,7 @@ import timeRange from './time-range';
 import serverBandwidth from './bandwidth';
 import editServiceName from './edit-name';
 import taskTracker from './task-tracker';
-import sshKey from './ssh-key';
+import sshKeySelector from './ssh-key-selector';
 
 export {
   inlinePropertyEditor,
@@ -14,7 +14,7 @@ export {
   serverBandwidth,
   editServiceName,
   taskTracker,
-  sshKey,
+  sshKeySelector,
 };
 
 export default { resourceSelector };
