+++ packages/manager/modules/manager-components/src/ssh-key-selector/index.js
@@ -2,15 +2,15 @@ import angular from 'angular';
 
 import '@ovh-ux/ui-kit';
 
-import component from './ssh-key.component';
-import service from './ssh-key.service';
+import component from './ssh-key-selector.component';
+import service from './ssh-key-selector.service';
 
 const moduleName = 'ovhManagerComponentsSshKey';
 
 angular
   .module(moduleName, ['oui'])
-  .component('sshKey', component)
-  .service('SshKeyService', service)
+  .component('sshKeySelector', component)
+  .service('SshKeySelectorService', service)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
