+++ packages/manager/modules/manager-components/src/ssh-key-selector/ssh-key-selector.service.js
@@ -1,4 +1,4 @@
-export default class SshKeyService {
+export default class SshKeySelectorService {
   /* @ngInject */
   constructor($http) {
     this.$http = $http;
