+++ packages/manager/modules/vps/src/vps/vps.module.js
@@ -13,7 +13,7 @@ import ngAtInternet from '@ovh-ux/ng-at-internet';
 import ngOvhCloudUniverseComponents from '@ovh-ux/ng-ovh-cloud-universe-components';
 import ngOvhDocUrl from '@ovh-ux/ng-ovh-doc-url';
 import ngOvhUserPref from '@ovh-ux/ng-ovh-user-pref';
-import { region, sshKey } from '@ovh-ux/manager-components';
+import { region, sshKeySelector } from '@ovh-ux/manager-components';
 
 import VpsTaskService from './vps-task.service';
 import VpsNotificationIpv6Service from '../import/notification.service';
@@ -69,7 +69,7 @@ angular
     ovhManagerVpsVeeam,
     ovhManagerVpsWindows,
     region,
-    sshKey,
+    sshKeySelector,
   ])
   .component(detailComponent.name, detailComponent)
   .component(headerComponent.name, headerComponent)
